% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Clean input CREDI response data}
\usage{
clean(input_df, mest_df, csv_wd, reverse_code, log_file)
}
\arguments{
\item{input_df}{Raw input data from the user-specified csv file.}

\item{mest_df}{Measurement parameter estimates.}

\item{csv_wd}{Working directory where input *.csv file is located.}

\item{reverse_code}{Logical. If TRUE, then reverse codes LF9, LF102, LFMH1, LFMH2, LFMH3, LFMH4, LFMH5, LFMH7, LFMH8, & LFMH9. If FALSE, then no reverse coding is applied.}

\item{log_file}{Name of the *.txt log file for the CREDI scoring.}
}
\description{
This calculates the posterior density function.
}
\examples{
clean(input_df, mest_df, csv_wd, reverse_code,log_file)      
}
\keyword{CREDI}
