% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lf_posterior_density.R
\name{lf_posterior_density}
\alias{lf_posterior_density}
\title{Posterior density for CREDI subscales}
\usage{
lf_posterior_density(THETAi, Yi, MUi, invS, TAU, LAMBDA, J, K)
}
\arguments{
\item{THETAi}{Size K vector of scaled scores for the i-th individual.}

\item{Yi}{Size Y vector of item responses for the i-th individual.}

\item{MUi}{Size K vector of prior's mean conditional on MONTHS.}

\item{invS}{Size K-by-K matrix of prior's inverse of var/cov matrix.}

\item{TAU}{Size J vector of item intercepts.}

\item{LAMBDA}{Size J-by-K matrix of item loadings.}

\item{J}{Number of items.}

\item{K}{Number of latent dimensions.}
}
\description{
This calculates the posterior density function.
}
\examples{
lf_posterior_density(THETAi, Yi, MUi, invS, TAU, LAMBDA,J,K)
}
\keyword{CREDI}
