% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_posterior_density.R
\name{sf_posterior_density}
\alias{sf_posterior_density}
\title{Posterior density for short form CREDI scale.}
\usage{
sf_posterior_density(
  THETAi,
  Yi,
  MUi,
  SIGMA_SQi,
  DELTA,
  ALPHA,
  J,
  weight = NULL
)
}
\arguments{
\item{THETAi}{Scalar representing scaled score for the i-th individual.}

\item{Yi}{Size N vector of item responses for the i-th individual.}

\item{MUi}{Scalar representing prior's mean conditional on AGE.}

\item{SIGMA_SQi}{Scalar representing prior variance conditional on AGE.}

\item{DELTA}{Size J vector of item difficulties.}

\item{ALPHA}{Size J vector of item discrimination values.}

\item{J}{Number of items.}

\item{weight}{Size J vector of weights for Bayesian Bootstrapping.}
}
\description{
This calculates the posterior density function ().
}
\examples{
sf_posterior_density(THETAi, Yi, MUi, SIGMA_SQi, DELTA, ALPHA,J)
}
\keyword{CREDI}
