% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_scores.R
\name{which_scores}
\alias{which_scores}
\title{Determine which scales should be scored given an individual's responses.}
\usage{
which_scores(Yi, mest_df = mest_df, min_items = min_items)
}
\arguments{
\item{Yi}{(numeric vector) with response data. Must be named according to CREDI_code.}

\item{mest_df}{(data.frame) Measurement parameter estimates.}

\item{min_items}{(integer) Minimum number of items that must have responses to score.}
}
\description{
This function determines whether there are a sufficient number of item responses
per scale in order to calculate a score
}
