% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Clean input CREDI response data}
\usage{
clean(input_df, mest_df, reverse_code, interactive, log)
}
\arguments{
\item{input_df}{(data.frame) Raw input data from the user-specified csv file.}

\item{mest_df}{(data.frame) Measurement parameter estimates.}

\item{reverse_code}{(logical). If TRUE, then reverse codes LF9, LF102, LFMH1, LFMH2, LFMH3, LFMH4, LFMH5, LFMH7, LFMH8, & LFMH9. If FALSE, then no reverse coding is applied.}

\item{log}{(string) Name of the *.txt log file for the CREDI scoring.}

\item{csv_wd}{(string) Working directory where input *.csv file is located.}
}
\description{
This calculates the posterior density function.
}
\examples{
clean(input_df, mest_df, reverse_code, interactive, log)
}
\keyword{CREDI}
